requirements.txt
numpy==1.12.1
opencv-python==3.4.0.12
python-telegram-bot==9.0.0